extract shared geometry utilities from bezier-parallel and ray-guided
